Move the unit tests for the payload base utils to a new file
Create a new file for these tests.
